Adiciona outros testes e ajusta impressões

+++ tests/test_entradas_saidas.sh
@@ -1,18 +1,23 @@
 #!/bin/bash
 result=0
 cp -r src/ tests/src_temp/
-sed -i 's/#define NREGSPORPAGINA 3/#define NREGSPORPAGINA 10/' tests/src_temp/types.h
 
-for input in tests/*.in; do
-  output=${input%.in}.out
-  rm -f arvore.dat
+for input in tests/teste_entradas_saidas/entradas/*; do
+  echo "Testando $input"
+  if [[ ${input: -1} == 'a' ]]; then
+    rm -f arvore.dat
+  fi
+
+  output=tests/teste_entradas_saidas/saidas/s${input##*/teste}
+  
   gcc -std=c99 -Wall -o main tests/src_temp/main.c
   ./main < $input > test.out
   diff -w -u1 --color=auto --palette='ad=1;3;38;5;154:de=1;3;38;5;9' test.out $output && echo "OK" || result=1
+
 done
 
-rm -f arvore.dat
 rm -f test.out
 rm -f main
 rm -rf tests/src_temp/
+
 exit $result
